# perf: speed up xqueue.upgrade() for large spaces