--- a/.luabench
+++ b/.luabench
@@ -1,0 +1,2 @@
+jit = 'on'
+duration = '10s'