Benchmark results

```
Tarantool version: Tarantool 3.1.0-0-g96f6d88
Tarantool build: Darwin-arm64-RelWithDebInfo (static)
Tarantool build flags:  -fexceptions -funwind-tables -fasynchronous-unwind-tables -fno-common  -fmacro-prefix-map=/var/folders/8x/1m5v3n6d4mn62g9w_65vvt_r0000gn/T/tarantool_install934702387=. -std=c11 -Wall -Wextra -Wno-gnu-alignof-expression -Wno-cast-function-type -O2 -g -DNDEBUG -ggdb -O2 
CPU: Apple M1 @ 8
JIT: Enabled
JIT: fold cse dce fwd dse narrow loop abc sink fuse
Duration: 3s

--- BENCH: 002_initial_stat_bench::bench_count
 3000000                24.54 ns/op       40744815 op/s        0 B/op   +9.90KB

--- BENCH: 002_initial_stat_bench::bench_iterate_all
 3000000               449.8 ns/op         2223177 op/s      117 B/op   +337.54MB
```

```
Tarantool version: Tarantool 3.1.0-0-g96f6d88
Tarantool build: Darwin-arm64-RelWithDebInfo (static)
Tarantool build flags:  -fexceptions -funwind-tables -fasynchronous-unwind-tables -fno-common  -fmacro-prefix-map=/var/folders/8x/1m5v3n6d4mn62g9w_65vvt_r0000gn/T/tarantool_install934702387=. -std=c11 -Wall -Wextra -Wno-gnu-alignof-expression -Wno-cast-function-type -O2 -g -DNDEBUG -ggdb -O2 
CPU: Apple M1 @ 8
JIT: Enabled
JIT: fold cse dce fwd dse narrow loop abc sink fuse
Duration: 3s

--- BENCH: 002_initial_stat_bench::bench_count
 4000000                24.82 ns/op       40296177 op/s        0 B/op   +9.70KB

--- BENCH: 002_initial_stat_bench::bench_iterate_all
 4000000               453.5 ns/op         2205092 op/s      117 B/op   +447.70MB
```

With replacing `:pairs({}, {iterator=box.index.ALL})` to several calls of `:count()` we speedup initial space.upgrade() for approximately 20x times.